+++ ArmyFit/Info.plist
@@ -1,29 +1,36 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
-<dict>
-	<key>NSPhotoLibraryUsageDescription</key>
-	<string>É necessário acesso as fotos para atualizar a foto de perfil e para poder compartilhar conteúdo no seu feed.</string>
-	<key>NSCameraUsageDescription</key>
-	<string>É necessário acesso a camera para atualizar a foto de perfil e para poder compartilhar conteúdo no seu feed.</string>
-	<key>UIApplicationSceneManifest</key>
-	<dict>
-		<key>UIApplicationSupportsMultipleScenes</key>
-		<false/>
-		<key>UISceneConfigurations</key>
-		<dict>
-			<key>UIWindowSceneSessionRoleApplication</key>
-			<array>
-				<dict>
-					<key>UISceneConfigurationName</key>
-					<string>Default Configuration</string>
-					<key>UISceneDelegateClassName</key>
-					<string>$(PRODUCT_MODULE_NAME).SceneDelegate</string>
-					<key>UISceneStoryboardFile</key>
-					<string>Main</string>
-				</dict>
-			</array>
-		</dict>
-	</dict>
-</dict>
+    <dict>
+        <key>NSPhotoLibraryUsageDescription</key>
+        <string>É necessário acesso as fotos para atualizar a foto de perfil e para poder compartilhar conteúdo no seu feed.</string>
+        <key>NSCameraUsageDescription</key>
+        <string>É necessário acesso a camera para atualizar a foto de perfil e para poder compartilhar conteúdo no seu feed.</string>
+        <key>UIApplicationSceneManifest</key>
+        <dict>
+            <key>UIApplicationSupportsMultipleScenes</key>
+            <false/>
+            <key>UISceneConfigurations</key>
+            <dict>
+                <key>UIWindowSceneSessionRoleApplication</key>
+                <array>
+                    <dict>
+                        <key>UISceneConfigurationName</key>
+                        <string>Default Configuration</string>
+                        <key>UISceneDelegateClassName</key>
+                        <string>$(PRODUCT_MODULE_NAME).SceneDelegate</string>
+                        <key>UISceneStoryboardFile</key>
+                        <string>Main</string>
+                    </dict>
+                </array>
+                <key>LSApplicationQueriesSchemes</key>
+                <array>
+                    <string>googlegmail</string>
+                    <string>ms-outlook</string>
+                    <string>readdle-spark</string>
+                    <string>ymail</string>
+                </array>
+            </dict>
+        </dict>
+    </dict>
 </plist>
