+++ ArmyFit/View/ProfileViewController.swift
@@ -111,29 +111,6 @@ class ProfileViewController: UIViewController {
         
         present(imagePicker, animated: true)
     }
-    
-    private func createEmailUrl(to: String, subject: String, body: String) -> URL? {
-        let subjectEncoded = subject.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed)!
-        let bodyEncoded = body.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed)!
-        
-        let gmailUrl = URL(string: "googlegmail://co?to=\(to)&subject=\(subjectEncoded)&body=\(bodyEncoded)")
-        let outlookUrl = URL(string: "ms-outlook://compose?to=\(to)&subject=\(subjectEncoded)")
-        let yahooMail = URL(string: "ymail://mail/compose?to=\(to)&subject=\(subjectEncoded)&body=\(bodyEncoded)")
-        let sparkUrl = URL(string: "readdle-spark://compose?recipient=\(to)&subject=\(subjectEncoded)&body=\(bodyEncoded)")
-        let defaultUrl = URL(string: "mailto:\(to)?subject=\(subjectEncoded)&body=\(bodyEncoded)")
-        
-        if let gmailUrl = gmailUrl, UIApplication.shared.canOpenURL(gmailUrl) {
-            return gmailUrl
-        } else if let outlookUrl = outlookUrl, UIApplication.shared.canOpenURL(outlookUrl) {
-            return outlookUrl
-        } else if let yahooMail = yahooMail, UIApplication.shared.canOpenURL(yahooMail) {
-            return yahooMail
-        } else if let sparkUrl = sparkUrl, UIApplication.shared.canOpenURL(sparkUrl) {
-            return sparkUrl
-        }
-        
-        return defaultUrl
-    }
 }
 
 //MARK: - PHPickerViewControllerDelegate
@@ -191,9 +168,8 @@ extension ProfileViewController: UITableViewDelegate {
         
         if indexPath.section == 1 {
             if indexPath.row == 0 {
-                let recipientEmail = "test@email.com"
-                let subject = "Multi client email support"
-                let body = "This code supports sending email via multiple different email apps on iOS! :)"
+                let recipientEmail = "vjeanlucas93@gmail.com"
+                let subject = "ArmyFit - Suporte ao cliente"
                 
                 // Show default mail composer
                 if MFMailComposeViewController.canSendMail() {
@@ -201,12 +177,11 @@ extension ProfileViewController: UITableViewDelegate {
                     mail.mailComposeDelegate = self
                     mail.setToRecipients([recipientEmail])
                     mail.setSubject(subject)
-                    mail.setMessageBody(body, isHTML: false)
                     
                     present(mail, animated: true)
                     
                     // Show third party email composer if default Mail app is not present
-                } else if let emailUrl = createEmailUrl(to: recipientEmail, subject: subject, body: body) {
+                } else if let emailUrl = viewModel.createEmailUrl(to: recipientEmail, subject: subject) {
                     UIApplication.shared.open(emailUrl)
                 }
             }
